Added AccountService

+++ src/main/java/com/mozzie/nbp/domain/models/Account.java
@@ -1,5 +1,6 @@
 package com.mozzie.nbp.domain.models;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +10,5 @@ public class Account {
     private String accountId;
     private String firstName;
     private String lastName;
-    private double balancePLN;
+    private BigDecimal balancePLN;
 }
